requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.37.35
botocore==1.37.35
certifi==2025.1.31
charset-normalizer==3.4.1
deepseek-ai==0.0.1
distro==1.9.0
Django==5.0.2
django-allauth==65.7.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
elevenlabs==1.57.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
openai==1.75.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
razorpay==1.4.2
redis==6.1.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
s3transfer==0.11.4
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
wheel==0.45.1
